SFA: Draw all 3 map block bitstreams
