+++ src/StarFoxAdventures/blocks.ts
@@ -7,10 +7,11 @@ import { nArray } from '../util';
 import * as GX from '../gx/gx_enum';
 import { GXMaterialBuilder } from "../gx/GXMaterialBuilder";
 
-import { ModelInstance, SFARenderer } from './render';
+import { SFARenderer } from './render';
 import { TextureCollection, SFATextureCollection, FakeTextureCollection } from './textures';
 import { getSubdir } from './resource';
 import { GameInfo } from './scenes';
+import { Shader, parseShader, SFA_SHADER_FIELDS, EARLY_SFA_SHADER_FIELDS, ModelInstance } from './models';
 
 export abstract class BlockFetcher {
     public abstract getBlock(mod: number, sub: number): ArrayBufferSlice | null;
@@ -123,7 +124,7 @@ export class SFABlockRenderer implements BlockRenderer {
         let offs = 0;
         const blockDv = blockData.createDataView();
 
-        let fields;
+        let fields: any;
         if (earlyFields) {
             fields = {
                 texOffset: 0x54,
@@ -142,6 +143,7 @@ export class SFABlockRenderer implements BlockRenderer {
                 jointCount: 0,
                 shaderOffset: 0x64,
                 shaderCount: 0xa0, // Polygon attributes and material information
+                shaderFields: EARLY_SFA_SHADER_FIELDS,
                 shaderSize: 0x40,
                 listOffset: 0x68,
                 listCount: 0x9f,
@@ -150,9 +152,9 @@ export class SFABlockRenderer implements BlockRenderer {
                 // numListBits: 6, // 6 is needed for mod12; 8 is needed for early crfort?!
                 numListBits: 8, // ??? should be 6 according to decompilation of demo????
                 numLayersOffset: 0x3b,
-                bitstreamOffset: 0x74, // Whoa...
+                bitsOffsets: [0x74], // Whoa...
                 // FIXME: There are three bitstreams, probably for opaque and transparent objects
-                bitstreamByteCount: 0x84,
+                bitsByteCounts: [0x84],
                 oldVat: true,
                 hasYTranslate: false,
                 oldShaders: true,
@@ -181,14 +183,13 @@ export class SFABlockRenderer implements BlockRenderer {
                     jointCount: 243,
                     shaderOffset: 56,
                     shaderCount: 248,
-                    shaderSize: 0x44,
+                    shaderFields: SFA_SHADER_FIELDS,
                     listOffset: 208,
                     listCount: 245,
                     listSize: 0x1c,
                     numListBits: 8,
-                    numLayersOffset: 0x41,
-                    bitstreamOffset: 212,
-                    bitstreamByteCount: 216,
+                    bitsOffsets: [212],
+                    bitsByteCounts: [216],
                     oldVat: false,
                     hasYTranslate: false,
                     oldShaders: false,
@@ -214,14 +215,13 @@ export class SFABlockRenderer implements BlockRenderer {
                     jointCount: 0,
                     shaderOffset: 0x64,
                     shaderCount: 0xa2,
-                    shaderSize: 0x44,
+                    shaderFields: SFA_SHADER_FIELDS,
                     listOffset: 0x68,
                     listCount: 0xa1, // TODO
                     listSize: 0x1c,
                     numListBits: 8,
-                    numLayersOffset: 0x41,
-                    bitstreamOffset: 0x78,
-                    bitstreamByteCount: 0x84,
+                    bitsOffsets: [0x78, 0x7c, 0x80],
+                    bitsByteCounts: [0x84, 0x86, 0x88],
                     oldVat: false,
                     hasYTranslate: true,
                     oldShaders: false,
@@ -280,79 +280,14 @@ export class SFABlockRenderer implements BlockRenderer {
         const shaderCount = blockDv.getUint8(fields.shaderCount);
         // console.log(`Loading ${polyCount} polytypes from 0x${polyOffset.toString(16)}`);
 
-        interface ShaderLayer {
-            texNum: number;
-            tevMode: number;
-        }
-
-        function parseShaderLayer(data: DataView): ShaderLayer {
-            return {
-                texNum: data.getUint32(0),
-                tevMode: data.getUint8(4),
-            };
-        }
-
-        interface Shader {
-            numLayers: number;
-            layers: ShaderLayer[],
-            hasTexCoord: boolean[];
-            enableCull: boolean;
-            flags: number;
-            hasTexmtx01: boolean;
-            hasTexmtx2: boolean;
-            attrFlags: number;
-        }
-
         const shaders: Shader[] = [];
         offs = shaderOffset;
-        enum ShaderFlags {
-            Cull = 0x8,
-        }
+        const shaderFields = SFA_SHADER_FIELDS;
         for (let i = 0; i < shaderCount; i++) {
-            const shader: Shader = {
-                numLayers: 0,
-                layers: [],
-                hasTexCoord: nArray(8, () => false),
-                enableCull: false,
-                flags: 0,
-                hasTexmtx01: false,
-                hasTexmtx2: false,
-                attrFlags: 0,
-            };
-
-            shader.numLayers = blockDv.getUint8(offs + fields.numLayersOffset);
-            if (shader.numLayers > 2) {
-                console.warn(`Number of shader layers greater than maximum (${shader.numLayers} / 2)`);
-                shader.numLayers = 2;
-            }
-            for (let i = 0; i < shader.numLayers; i++) {
-                const layer = parseShaderLayer(blockData.subarray(offs + 0x24 + i * 8).createDataView());
-                shader.layers.push(layer);
-            }
-
-            for (let j = 0; j < shader.numLayers; j++) {
-                shader.hasTexCoord[j] = true;
-            }
-            shader.flags = blockDv.getUint32(offs + 0x3c);
-            // FIXME: find this field's offset for demo files
-            if (fields.oldShaders) {
-                // FIXME: this is from decompilation but it doesn't seem to work in cloudtreasure...
-                // shader.enableCull = (blockDv.getUint8(offs + 0x38) & 0x4) != 0;
-                shader.enableCull = true;
-            } else {
-                shader.enableCull = (shader.flags & ShaderFlags.Cull) != 0;
-            }
-
-            // FIXME: the texmtx stuff below is broken or not present in SFA...
-            // shader.hasTexmtx01 = blockDv.getUint32(offs + 8) == 1 || blockDv.getUint32(offs + 20) == 1;
-            // shader.hasTexmtx2 = (blockDv.getUint32(offs + 64 + 2) & 0x80) != 0;
-            shader.hasTexmtx01 = blockDv.getUint32(offs + 0x34) != 0;
-            shader.hasTexmtx2 = false;
-
-            shader.attrFlags = blockDv.getUint8(offs + 0x40);
-            
+            const shaderBin = blockData.subarray(offs, shaderFields.size).createDataView();
+            const shader = parseShader(shaderBin, shaderFields);
             shaders.push(shader);
-            offs += fields.shaderSize;
+            offs += shaderFields.size;
         }
 
         const vcd: GX_VtxDesc[] = [];
@@ -420,9 +355,12 @@ export class SFABlockRenderer implements BlockRenderer {
         const listCount = blockDv.getUint8(fields.listCount);
         // console.log(`Loading ${chunkCount} display lists from 0x${chunkOffset.toString(16)}`);
 
-        const bitstreamOffset = blockDv.getUint32(fields.bitstreamOffset);
-        const bitstreamByteCount = blockDv.getUint16(fields.bitstreamByteCount);
-        // console.log(`Loading ${bitsCount} bits from 0x${bitsOffset.toString(16)}`);
+        const bitsOffsets = [];
+        const bitsByteCounts = [];
+        for (let i = 0; i < fields.bitsOffsets.length; i++) {
+            bitsOffsets.push(blockDv.getUint32(fields.bitsOffsets[i]));
+            bitsByteCounts.push(blockDv.getUint16(fields.bitsByteCounts[i]));
+        }
 
         let texMtxCount = 0;
         if (fields.hasJoints) {
@@ -435,287 +373,297 @@ export class SFABlockRenderer implements BlockRenderer {
             this.yTranslate = 0;
         }
 
-        const bits = new LowBitReader(blockDv, bitstreamOffset);
-        let done = false;
-        let curShader = shaders[0];
-        while (!done) {
-            const opcode = bits.get(4);
-            switch (opcode) {
-            case 1: // Set polygon type
-                curShader = shaders[bits.get(6)];
-                break;
-            case 2: // Geometry
-                const listNum = bits.get(fields.numListBits);
-                // console.log(`Drawing display list #${chunkNum}`);
-                if (listNum >= listCount) {
-                    console.warn(`Can't draw display list #${listNum} (out of range)`);
-                    continue;
-                }
-                offs = listOffset + listNum * fields.listSize;
-                const dlOffset = blockDv.getUint32(offs);
-                const dlSize = blockDv.getUint16(offs + 4);
-                const displayList = blockData.subarray(dlOffset, dlSize);
-
-                const vtxArrays: GX_Array[] = [];
-                vtxArrays[GX.Attr.POS] = { buffer: vertBuffer, offs: 0, stride: 6 /*getAttributeByteSize(vat[0], GX.Attr.POS)*/ };
-                if (fields.hasNormals) {
-                    vtxArrays[GX.Attr.NRM] = { buffer: nrmBuffer, offs: 0, stride: (nrmTypeFlags & 8) != 0 ? 9 : 3 /*getAttributeByteSize(vat[0], GX.Attr.NRM)*/ };
-                }
-                vtxArrays[GX.Attr.CLR0] = { buffer: clrBuffer, offs: 0, stride: 2 /*getAttributeByteSize(vat[0], GX.Attr.CLR0)*/ };
-                for (let t = 0; t < 8; t++) {
-                    vtxArrays[GX.Attr.TEX0 + t] = { buffer: texcoordBuffer, offs: 0, stride: 4 /*getAttributeByteSize(vat[0], GX.Attr.TEX0)*/ };
-                }
-
-                try {
-                    const newModel = new ModelInstance(vtxArrays, vcd, vat, displayList);
-
-                    const mb = new GXMaterialBuilder('Basic');
-                    if ((curShader.flags & 0x40000000) || (curShader.flags & 0x20000000)) {
-                        mb.setBlendMode(GX.BlendMode.BLEND, GX.BlendFactor.SRCALPHA, GX.BlendFactor.INVSRCALPHA, GX.LogicOp.NOOP);
-                        mb.setZMode(true, GX.CompareType.LEQUAL, false);
-                        mb.setAlphaCompare(GX.CompareType.ALWAYS, 0, GX.AlphaOp.AND, GX.CompareType.ALWAYS, 0);
-                    } else {
-                        mb.setBlendMode(GX.BlendMode.NONE, GX.BlendFactor.ONE, GX.BlendFactor.ZERO, GX.LogicOp.NOOP);
-                        mb.setZMode(true, GX.CompareType.LEQUAL, true);
-                        if (((curShader.flags & 0x400) == 0) || ((curShader.flags & 0x80) != 0)) {
+        const self = this;
+        function runBitstream(bitsOffset: number) {
+            const bits = new LowBitReader(blockDv, bitsOffset);
+            let done = false;
+            let curShader = shaders[0];
+            while (!done) {
+                const opcode = bits.get(4);
+                switch (opcode) {
+                case 1: // Set polygon type
+                    curShader = shaders[bits.get(6)];
+                    break;
+                case 2: // Geometry
+                    const listNum = bits.get(fields.numListBits);
+                    // console.log(`Drawing display list #${chunkNum}`);
+                    if (listNum >= listCount) {
+                        console.warn(`Can't draw display list #${listNum} (out of range)`);
+                        continue;
+                    }
+                    offs = listOffset + listNum * fields.listSize;
+                    const dlOffset = blockDv.getUint32(offs);
+                    const dlSize = blockDv.getUint16(offs + 4);
+                    const displayList = blockData.subarray(dlOffset, dlSize);
+    
+                    const vtxArrays: GX_Array[] = [];
+                    vtxArrays[GX.Attr.POS] = { buffer: vertBuffer, offs: 0, stride: 6 /*getAttributeByteSize(vat[0], GX.Attr.POS)*/ };
+                    if (fields.hasNormals) {
+                        vtxArrays[GX.Attr.NRM] = { buffer: nrmBuffer, offs: 0, stride: (nrmTypeFlags & 8) != 0 ? 9 : 3 /*getAttributeByteSize(vat[0], GX.Attr.NRM)*/ };
+                    }
+                    vtxArrays[GX.Attr.CLR0] = { buffer: clrBuffer, offs: 0, stride: 2 /*getAttributeByteSize(vat[0], GX.Attr.CLR0)*/ };
+                    for (let t = 0; t < 8; t++) {
+                        vtxArrays[GX.Attr.TEX0 + t] = { buffer: texcoordBuffer, offs: 0, stride: 4 /*getAttributeByteSize(vat[0], GX.Attr.TEX0)*/ };
+                    }
+    
+                    try {
+                        const newModel = new ModelInstance(vtxArrays, vcd, vat, displayList);
+    
+                        const mb = new GXMaterialBuilder('Basic');
+                        if ((curShader.flags & 0x40000000) || (curShader.flags & 0x20000000)) {
+                            mb.setBlendMode(GX.BlendMode.BLEND, GX.BlendFactor.SRCALPHA, GX.BlendFactor.INVSRCALPHA, GX.LogicOp.NOOP);
+                            mb.setZMode(true, GX.CompareType.LEQUAL, false);
                             mb.setAlphaCompare(GX.CompareType.ALWAYS, 0, GX.AlphaOp.AND, GX.CompareType.ALWAYS, 0);
                         } else {
-                            mb.setAlphaCompare(GX.CompareType.GREATER, 0, GX.AlphaOp.AND, GX.CompareType.GREATER, 0);
-                        }
-                    }
-                    mb.setChanCtrl(GX.ColorChannelID.COLOR0A0, false, GX.ColorSrc.REG, GX.ColorSrc.VTX, 0, GX.DiffuseFunction.NONE, GX.AttenuationFunction.NONE);
-                    mb.setCullMode(curShader.enableCull ? GX.CullMode.BACK : GX.CullMode.NONE);
-
-                    let tevStage = 0;
-                    let texcoordId = GX.TexCoordID.TEXCOORD0;
-                    let texmapId = GX.TexMapID.TEXMAP0;
-                    let texGenSrc = GX.TexGenSrc.TEX0;
-                    let cprevIsValid = false;
-                    let aprevIsValid = false;
-                    function addTevStagesForTextureWithSkyAmbient() {
-                        // TODO: set texture matrix
-                        mb.setTexCoordGen(texcoordId, GX.TexGenType.MTX2x4, texGenSrc, GX.TexGenMatrix.IDENTITY);
-
-                        // mb.setTevKColor (does not exist)
-                        // TODO: The game multiplies by a sky-related ambient color
-                        // mb.setTevKColorSel(tevStage, GX.KonstColorSel.KCSEL_K0);
-                        // Stage 1: Multiply vertex color by ambient sky color
-                        mb.setTevDirect(tevStage);
-                        mb.setTevOrder(tevStage, GX.TexCoordID.TEXCOORD_NULL, GX.TexMapID.TEXMAP_NULL, GX.RasColorChannelID.COLOR0A0);
-                        mb.setTevColorIn(tevStage, GX.CC.ZERO, GX.CC.ONE /*GX.CombineColorInput.KONST*/, GX.CC.RASC, GX.CC.ZERO);
-                        mb.setTevAlphaIn(tevStage, GX.CA.ZERO, GX.CA.ZERO, GX.CA.ZERO, GX.CA.ZERO);
-                        mb.setTevColorOp(tevStage, GX.TevOp.ADD, GX.TevBias.ZERO, GX.TevScale.SCALE_1, true, GX.Register.PREV);
-                        mb.setTevAlphaOp(tevStage, GX.TevOp.ADD, GX.TevBias.ZERO, GX.TevScale.SCALE_1, true, GX.Register.PREV);
-
-                        // Stage 2: Blend previous stage with vertex color by vertex alpha
-                        mb.setTevDirect(tevStage + 1);
-                        mb.setTevOrder(tevStage + 1, GX.TexCoordID.TEXCOORD_NULL, GX.TexMapID.TEXMAP_NULL, GX.RasColorChannelID.COLOR0A0);
-                        mb.setTevColorIn(tevStage + 1, GX.CC.CPREV, GX.CC.RASC, GX.CC.RASA, GX.CC.ZERO);
-                        mb.setTevAlphaIn(tevStage + 1, GX.CA.ZERO, GX.CA.ZERO, GX.CA.ZERO, GX.CA.ZERO);
-                        mb.setTevColorOp(tevStage + 1, GX.TevOp.ADD, GX.TevBias.ZERO, GX.TevScale.SCALE_1, true, GX.Register.PREV);
-                        mb.setTevAlphaOp(tevStage + 1, GX.TevOp.ADD, GX.TevBias.ZERO, GX.TevScale.SCALE_1, true, GX.Register.PREV);
-
-                        // Stage 3: Multiply by texture
-                        mb.setTevDirect(tevStage + 2);
-                        mb.setTevOrder(tevStage + 2, texcoordId, texmapId, GX.RasColorChannelID.COLOR_ZERO /* GX_COLOR_NULL */);
-                        mb.setTevColorIn(tevStage + 2, GX.CC.ZERO, GX.CC.CPREV, GX.CC.TEXC, GX.CC.ZERO);
-                        mb.setTevAlphaIn(tevStage + 2, GX.CA.ZERO, GX.CA.ZERO, GX.CA.ZERO, GX.CA.TEXA);
-                        mb.setTevColorOp(tevStage + 2, GX.TevOp.ADD, GX.TevBias.ZERO, GX.TevScale.SCALE_1, true, GX.Register.PREV);
-                        mb.setTevAlphaOp(tevStage + 2, GX.TevOp.ADD, GX.TevBias.ZERO, GX.TevScale.SCALE_1, true, GX.Register.PREV);
-
-                        tevStage += 3;
-                        texcoordId++;
-                        texmapId++;
-                        texGenSrc++;
-                    }
-
-                    function addTevStagesForTextureWithMode(mode: number) {
-                        // TODO: set texture matrix
-                        mb.setTexCoordGen(texcoordId, GX.TexGenType.MTX2x4, texGenSrc, GX.TexGenMatrix.IDENTITY);
-
-                        mb.setTevDirect(tevStage);
-                        mb.setTevOrder(tevStage, texcoordId, texmapId, GX.RasColorChannelID.COLOR0A0);
-                        // Only modes 0 and 9 occur in map blocks. Other modes
-                        // occur in object and character models.
-                        switch (mode) {
-                        case 0:
-                            mb.setTevColorIn(tevStage, GX.CC.ZERO, GX.CC.TEXC, GX.CC.RASC, GX.CC.ZERO);
-                            break;
-                        case 9:
-                            mb.setTevColorIn(tevStage, GX.CC.TEXC, GX.CC.CPREV, GX.CC.APREV, GX.CC.ZERO);
-                            break;
-                        default:
-                            console.warn(`Unhandled tev color-in mode ${mode}`);
-                            break;
-                        }
-
-                        if (!aprevIsValid) {
-                            mb.setTevAlphaIn(tevStage, GX.CA.ZERO, GX.CA.TEXA, GX.CA.RASA, GX.CA.ZERO);
-                            aprevIsValid = true;
-                        } else {
-                            mb.setTevAlphaIn(tevStage, GX.CA.ZERO, GX.CA.TEXA, GX.CA.APREV, GX.CA.ZERO);
-                        }
-                        mb.setTevColorOp(tevStage, GX.TevOp.ADD, GX.TevBias.ZERO, GX.TevScale.SCALE_1, true, GX.Register.PREV);
-                        mb.setTevAlphaOp(tevStage, GX.TevOp.ADD, GX.TevBias.ZERO, GX.TevScale.SCALE_1, true, GX.Register.PREV);
-                        cprevIsValid = true;
-
-                        tevStage++;
-                        texcoordId++;
-                        texmapId++;
-                        texGenSrc++;
-                    }
-
-                    function addTevStageForTextureWithWhiteKonst(colorInMode: number) {
-                        // TODO: handle color. map block renderer always passes opaque white to this function.
-                        
-                        // TODO: set texture matrix
-                        mb.setTexCoordGen(texcoordId, GX.TexGenType.MTX2x4, texGenSrc, GX.TexGenMatrix.IDENTITY);
-
-                        mb.setTevDirect(tevStage);
-                        mb.setTevOrder(tevStage, texcoordId, texmapId, GX.RasColorChannelID.COLOR0A0);
-                        switch (colorInMode) {
-                        case 0:
-                            mb.setTevColorIn(tevStage, GX.CC.ZERO, GX.CC.TEXC, GX.CC.ONE /* GX.CC.KONST */, GX.CC.ZERO);
-                            break;
-                        default:
-                            console.warn(`Unhandled colorInMode ${colorInMode}`);
-                            break;
-                        }
-
-                        if (!aprevIsValid) {
-                            mb.setTevAlphaIn(tevStage, GX.CA.ZERO, GX.CA.TEXA, GX.CA.RASA, GX.CA.ZERO);
-                            aprevIsValid = true;
-                        } else {
-                            mb.setTevAlphaIn(tevStage, GX.CA.ZERO, GX.CA.TEXA, GX.CA.APREV, GX.CA.ZERO);
-                        }
-                        mb.setTevColorOp(tevStage, GX.TevOp.ADD, GX.TevBias.ZERO, GX.TevScale.SCALE_1, true, GX.Register.PREV);
-                        mb.setTevAlphaOp(tevStage, GX.TevOp.ADD, GX.TevBias.ZERO, GX.TevScale.SCALE_1, true, GX.Register.PREV);
-                        cprevIsValid = true;
-
-                        tevStage++;
-                        texcoordId++;
-                        texmapId++;
-                        texGenSrc++;
-                    }
-
-                    function addTevStageForMultVtxColor() {
-                        // TODO: handle konst alpha. map block renderer always passes opaque white to this function.
-
-                        mb.setTevDirect(tevStage);
-                        mb.setTevOrder(tevStage, GX.TexCoordID.TEXCOORD_NULL, GX.TexMapID.TEXMAP_NULL, GX.RasColorChannelID.COLOR0A0);
-                        mb.setTevKAlphaSel(tevStage, GX.KonstAlphaSel.KASEL_1); // TODO: handle non-opaque alpha
-                        if (tevStage === 0 || !cprevIsValid) {
-                            mb.setTevColorIn(tevStage, GX.CC.ZERO, GX.CC.ZERO, GX.CC.ZERO, GX.CC.RASC);
-                            mb.setTevAlphaIn(tevStage, GX.CA.ZERO, GX.CA.ZERO, GX.CA.ZERO, GX.CA.KONST);
-                        } else {
-                            mb.setTevColorIn(tevStage, GX.CC.ZERO, GX.CC.CPREV, GX.CC.RASC, GX.CC.ZERO);
-                            mb.setTevAlphaIn(tevStage, GX.CA.ZERO, GX.CA.APREV, GX.CA.KONST, GX.CA.ZERO);
-                        }
-                        mb.setTevColorOp(tevStage, GX.TevOp.ADD, GX.TevBias.ZERO, GX.TevScale.SCALE_1, true, GX.Register.PREV);
-                        mb.setTevAlphaOp(tevStage, GX.TevOp.ADD, GX.TevBias.ZERO, GX.TevScale.SCALE_1, true, GX.Register.PREV);
-                        cprevIsValid = true;
-                        tevStage++;
-                    }
-
-                    if (curShader.numLayers === 2 && (curShader.layers[1].tevMode & 0x7f) === 9) {
-                        addTevStageForTextureWithWhiteKonst(0);
-                        addTevStagesForTextureWithMode(9);
-                        addTevStageForMultVtxColor();
-                    } else {
-                        for (let i = 0; i < curShader.numLayers; i++) {
-                            const layer = curShader.layers[i];
-                            if (curShader.flags & 0x40000) {
-                                addTevStagesForTextureWithSkyAmbient();
+                            mb.setBlendMode(GX.BlendMode.NONE, GX.BlendFactor.ONE, GX.BlendFactor.ZERO, GX.LogicOp.NOOP);
+                            mb.setZMode(true, GX.CompareType.LEQUAL, true);
+                            if (((curShader.flags & 0x400) == 0) || ((curShader.flags & 0x80) != 0)) {
+                                mb.setAlphaCompare(GX.CompareType.ALWAYS, 0, GX.AlphaOp.AND, GX.CompareType.ALWAYS, 0);
                             } else {
-                                addTevStagesForTextureWithMode(layer.tevMode & 0x7f);
+                                mb.setAlphaCompare(GX.CompareType.GREATER, 0, GX.AlphaOp.AND, GX.CompareType.GREATER, 0);
+                            }
+                        }
+                        mb.setChanCtrl(GX.ColorChannelID.COLOR0A0, false, GX.ColorSrc.REG, GX.ColorSrc.VTX, 0, GX.DiffuseFunction.NONE, GX.AttenuationFunction.NONE);
+                        mb.setCullMode(curShader.enableCull ? GX.CullMode.BACK : GX.CullMode.NONE);
+    
+                        let tevStage = 0;
+                        let texcoordId = GX.TexCoordID.TEXCOORD0;
+                        let texmapId = GX.TexMapID.TEXMAP0;
+                        let texGenSrc = GX.TexGenSrc.TEX0;
+                        let cprevIsValid = false;
+                        let aprevIsValid = false;
+                        function addTevStagesForTextureWithSkyAmbient() {
+                            // TODO: set texture matrix
+                            mb.setTexCoordGen(texcoordId, GX.TexGenType.MTX2x4, texGenSrc, GX.TexGenMatrix.IDENTITY);
+    
+                            // mb.setTevKColor (does not exist)
+                            // TODO: The game multiplies by a sky-related ambient color
+                            // mb.setTevKColorSel(tevStage, GX.KonstColorSel.KCSEL_K0);
+                            // Stage 1: Multiply vertex color by ambient sky color
+                            mb.setTevDirect(tevStage);
+                            mb.setTevOrder(tevStage, GX.TexCoordID.TEXCOORD_NULL, GX.TexMapID.TEXMAP_NULL, GX.RasColorChannelID.COLOR0A0);
+                            mb.setTevColorIn(tevStage, GX.CC.ZERO, GX.CC.ONE /*GX.CombineColorInput.KONST*/, GX.CC.RASC, GX.CC.ZERO);
+                            mb.setTevAlphaIn(tevStage, GX.CA.ZERO, GX.CA.ZERO, GX.CA.ZERO, GX.CA.ZERO);
+                            mb.setTevColorOp(tevStage, GX.TevOp.ADD, GX.TevBias.ZERO, GX.TevScale.SCALE_1, true, GX.Register.PREV);
+                            mb.setTevAlphaOp(tevStage, GX.TevOp.ADD, GX.TevBias.ZERO, GX.TevScale.SCALE_1, true, GX.Register.PREV);
+    
+                            // Stage 2: Blend previous stage with vertex color by vertex alpha
+                            mb.setTevDirect(tevStage + 1);
+                            mb.setTevOrder(tevStage + 1, GX.TexCoordID.TEXCOORD_NULL, GX.TexMapID.TEXMAP_NULL, GX.RasColorChannelID.COLOR0A0);
+                            mb.setTevColorIn(tevStage + 1, GX.CC.CPREV, GX.CC.RASC, GX.CC.RASA, GX.CC.ZERO);
+                            mb.setTevAlphaIn(tevStage + 1, GX.CA.ZERO, GX.CA.ZERO, GX.CA.ZERO, GX.CA.ZERO);
+                            mb.setTevColorOp(tevStage + 1, GX.TevOp.ADD, GX.TevBias.ZERO, GX.TevScale.SCALE_1, true, GX.Register.PREV);
+                            mb.setTevAlphaOp(tevStage + 1, GX.TevOp.ADD, GX.TevBias.ZERO, GX.TevScale.SCALE_1, true, GX.Register.PREV);
+    
+                            // Stage 3: Multiply by texture
+                            mb.setTevDirect(tevStage + 2);
+                            mb.setTevOrder(tevStage + 2, texcoordId, texmapId, GX.RasColorChannelID.COLOR_ZERO /* GX_COLOR_NULL */);
+                            mb.setTevColorIn(tevStage + 2, GX.CC.ZERO, GX.CC.CPREV, GX.CC.TEXC, GX.CC.ZERO);
+                            mb.setTevAlphaIn(tevStage + 2, GX.CA.ZERO, GX.CA.ZERO, GX.CA.ZERO, GX.CA.TEXA);
+                            mb.setTevColorOp(tevStage + 2, GX.TevOp.ADD, GX.TevBias.ZERO, GX.TevScale.SCALE_1, true, GX.Register.PREV);
+                            mb.setTevAlphaOp(tevStage + 2, GX.TevOp.ADD, GX.TevBias.ZERO, GX.TevScale.SCALE_1, true, GX.Register.PREV);
+    
+                            tevStage += 3;
+                            texcoordId++;
+                            texmapId++;
+                            texGenSrc++;
+                        }
+    
+                        function addTevStagesForTextureWithMode(mode: number) {
+                            // TODO: set texture matrix
+                            mb.setTexCoordGen(texcoordId, GX.TexGenType.MTX2x4, texGenSrc, GX.TexGenMatrix.IDENTITY);
+    
+                            mb.setTevDirect(tevStage);
+                            mb.setTevOrder(tevStage, texcoordId, texmapId, GX.RasColorChannelID.COLOR0A0);
+                            // Only modes 0 and 9 occur in map blocks. Other modes
+                            // occur in object and character models.
+                            switch (mode) {
+                            case 0:
+                                mb.setTevColorIn(tevStage, GX.CC.ZERO, GX.CC.TEXC, GX.CC.RASC, GX.CC.ZERO);
+                                break;
+                            case 9:
+                                mb.setTevColorIn(tevStage, GX.CC.TEXC, GX.CC.CPREV, GX.CC.APREV, GX.CC.ZERO);
+                                break;
+                            default:
+                                console.warn(`Unhandled tev color-in mode ${mode}`);
+                                break;
+                            }
+    
+                            if (!aprevIsValid) {
+                                mb.setTevAlphaIn(tevStage, GX.CA.ZERO, GX.CA.TEXA, GX.CA.RASA, GX.CA.ZERO);
+                                aprevIsValid = true;
+                            } else {
+                                mb.setTevAlphaIn(tevStage, GX.CA.ZERO, GX.CA.TEXA, GX.CA.APREV, GX.CA.ZERO);
+                            }
+                            mb.setTevColorOp(tevStage, GX.TevOp.ADD, GX.TevBias.ZERO, GX.TevScale.SCALE_1, true, GX.Register.PREV);
+                            mb.setTevAlphaOp(tevStage, GX.TevOp.ADD, GX.TevBias.ZERO, GX.TevScale.SCALE_1, true, GX.Register.PREV);
+                            cprevIsValid = true;
+    
+                            tevStage++;
+                            texcoordId++;
+                            texmapId++;
+                            texGenSrc++;
+                        }
+    
+                        function addTevStageForTextureWithWhiteKonst(colorInMode: number) {
+                            // TODO: handle color. map block renderer always passes opaque white to this function.
+                            
+                            // TODO: set texture matrix
+                            mb.setTexCoordGen(texcoordId, GX.TexGenType.MTX2x4, texGenSrc, GX.TexGenMatrix.IDENTITY);
+    
+                            mb.setTevDirect(tevStage);
+                            mb.setTevOrder(tevStage, texcoordId, texmapId, GX.RasColorChannelID.COLOR0A0);
+                            switch (colorInMode) {
+                            case 0:
+                                mb.setTevColorIn(tevStage, GX.CC.ZERO, GX.CC.TEXC, GX.CC.ONE /* GX.CC.KONST */, GX.CC.ZERO);
+                                break;
+                            default:
+                                console.warn(`Unhandled colorInMode ${colorInMode}`);
+                                break;
+                            }
+    
+                            if (!aprevIsValid) {
+                                mb.setTevAlphaIn(tevStage, GX.CA.ZERO, GX.CA.TEXA, GX.CA.RASA, GX.CA.ZERO);
+                                aprevIsValid = true;
+                            } else {
+                                mb.setTevAlphaIn(tevStage, GX.CA.ZERO, GX.CA.TEXA, GX.CA.APREV, GX.CA.ZERO);
+                            }
+                            mb.setTevColorOp(tevStage, GX.TevOp.ADD, GX.TevBias.ZERO, GX.TevScale.SCALE_1, true, GX.Register.PREV);
+                            mb.setTevAlphaOp(tevStage, GX.TevOp.ADD, GX.TevBias.ZERO, GX.TevScale.SCALE_1, true, GX.Register.PREV);
+                            cprevIsValid = true;
+    
+                            tevStage++;
+                            texcoordId++;
+                            texmapId++;
+                            texGenSrc++;
+                        }
+    
+                        function addTevStageForMultVtxColor() {
+                            // TODO: handle konst alpha. map block renderer always passes opaque white to this function.
+    
+                            mb.setTevDirect(tevStage);
+                            mb.setTevOrder(tevStage, GX.TexCoordID.TEXCOORD_NULL, GX.TexMapID.TEXMAP_NULL, GX.RasColorChannelID.COLOR0A0);
+                            mb.setTevKAlphaSel(tevStage, GX.KonstAlphaSel.KASEL_1); // TODO: handle non-opaque alpha
+                            if (tevStage === 0 || !cprevIsValid) {
+                                mb.setTevColorIn(tevStage, GX.CC.ZERO, GX.CC.ZERO, GX.CC.ZERO, GX.CC.RASC);
+                                mb.setTevAlphaIn(tevStage, GX.CA.ZERO, GX.CA.ZERO, GX.CA.ZERO, GX.CA.KONST);
+                            } else {
+                                mb.setTevColorIn(tevStage, GX.CC.ZERO, GX.CC.CPREV, GX.CC.RASC, GX.CC.ZERO);
+                                mb.setTevAlphaIn(tevStage, GX.CA.ZERO, GX.CA.APREV, GX.CA.KONST, GX.CA.ZERO);
+                            }
+                            mb.setTevColorOp(tevStage, GX.TevOp.ADD, GX.TevBias.ZERO, GX.TevScale.SCALE_1, true, GX.Register.PREV);
+                            mb.setTevAlphaOp(tevStage, GX.TevOp.ADD, GX.TevBias.ZERO, GX.TevScale.SCALE_1, true, GX.Register.PREV);
+                            cprevIsValid = true;
+                            tevStage++;
+                        }
+    
+                        if (curShader.layers.length === 2 && (curShader.layers[1].tevMode & 0x7f) === 9) {
+                            addTevStageForTextureWithWhiteKonst(0);
+                            addTevStagesForTextureWithMode(9);
+                            addTevStageForMultVtxColor();
+                        } else {
+                            for (let i = 0; i < curShader.layers.length; i++) {
+                                const layer = curShader.layers[i];
+                                if (curShader.flags & 0x40000) {
+                                    addTevStagesForTextureWithSkyAmbient();
+                                } else {
+                                    addTevStagesForTextureWithMode(layer.tevMode & 0x7f);
+                                }
+                            }
+                        }
+    
+                        newModel.setMaterial(mb.finish());
+    
+                        const textures = [];
+                        for (let i = 0; i < curShader.layers.length; i++) {
+                            textures.push(texColl.getTexture(device, texIds[curShader.layers[i].texNum], true));
+                        }
+                        newModel.setTextures(textures);
+    
+                        self.models.push(newModel);
+                    } catch (e) {
+                        console.error(e);
+                    }
+                    break;
+                case 3: // Set vertex attributes
+                    vcd[GX.Attr.PNMTXIDX].type = GX.AttrType.NONE;
+                    for (let i = 0; i < 8; i++) {
+                        vcd[GX.Attr.TEX0MTXIDX + i].type = GX.AttrType.NONE;
+                    }
+    
+                    if (fields.hasJoints && matrixIdxThing >= 2) {
+                        vcd[GX.Attr.PNMTXIDX].type = GX.AttrType.DIRECT;
+                        let texmtxNum = 0;
+                        // FIXME: what is this?
+                        // if (shaders[curShader].hasTexmtx01) {
+                        //     vcd[GX.Attr.TEX0MTXIDX + texmtxNum].type = GX.AttrType.DIRECT; 
+                        //     texmtxNum++;
+                        //     vcd[GX.Attr.TEX0MTXIDX + texmtxNum].type = GX.AttrType.DIRECT;
+                        //     texmtxNum++;
+                        // }
+                        vcd[GX.Attr.TEX0MTXIDX + texmtxNum].type = GX.AttrType.DIRECT;
+                        texmtxNum++;
+                        for (let i = 0; i < texMtxCount; i++) {
+                            vcd[GX.Attr.TEX0MTXIDX + texmtxNum].type = GX.AttrType.DIRECT;
+                            texmtxNum++;
+                        }
+                    }
+    
+                    const posDesc = bits.get(1);
+                    vcd[GX.Attr.POS].type = posDesc ? GX.AttrType.INDEX16 : GX.AttrType.INDEX8;
+    
+                    if (fields.hasNormals && (curShader.attrFlags & 1) != 0) {
+                        const nrmDesc = bits.get(1);
+                        if ((nrmTypeFlags & 8) != 0) {
+                            vcd[GX.Attr.NRM].type = GX.AttrType.NONE;
+                            vcd[GX.Attr.NBT].type = nrmDesc ? GX.AttrType.INDEX16 : GX.AttrType.INDEX8;
+                        } else {
+                            vcd[GX.Attr.NBT].type = GX.AttrType.NONE;
+                            vcd[GX.Attr.NRM].type = nrmDesc ? GX.AttrType.INDEX16 : GX.AttrType.INDEX8;
+                        }
+                    } else {
+                        vcd[GX.Attr.NRM].type = GX.AttrType.NONE;
+                    }
+    
+                    if ((curShader.attrFlags & 2) != 0) {
+                        const clr0Desc = bits.get(1);
+                        vcd[GX.Attr.CLR0].type = clr0Desc ? GX.AttrType.INDEX16 : GX.AttrType.INDEX8;
+                    } else {
+                        vcd[GX.Attr.CLR0].type = GX.AttrType.NONE;
+                    }
+    
+                    const texCoordDesc = bits.get(1);
+                    if (curShader.layers.length > 0) {
+                        // Note: texCoordDesc applies to all texture coordinates in the vertex
+                        for (let t = 0; t < 8; t++) {
+                            if (t < curShader.layers.length) {
+                                vcd[GX.Attr.TEX0 + t].type = texCoordDesc ? GX.AttrType.INDEX16 : GX.AttrType.INDEX8;
+                            } else {
+                                vcd[GX.Attr.TEX0 + t].type = GX.AttrType.NONE;
                             }
                         }
                     }
-
-                    newModel.setMaterial(mb.finish());
-
-                    const textures = [];
-                    for (let i = 0; i < curShader.numLayers; i++) {
-                        textures.push(texColl.getTexture(device, texIds[curShader.layers[i].texNum], true));
+                    break;
+                case 4: // Set weights (skipped by SFA block renderer)
+                    const numWeights = bits.get(4);
+                    for (let i = 0; i < numWeights; i++) {
+                        bits.get(8);
                     }
-                    newModel.setTextures(textures);
+                    break;
+                case 5: // End
+                    done = true;
+                    break;
+                default:
+                    console.warn(`Skipping unknown model bits opcode ${opcode}`);
+                    break;
+                }
+            }
+        }
 
-                    this.models.push(newModel);
-                } catch (e) {
-                    console.error(e);
-                }
-                break;
-            case 3: // Set vertex attributes
-                vcd[GX.Attr.PNMTXIDX].type = GX.AttrType.NONE;
-                for (let i = 0; i < 8; i++) {
-                    vcd[GX.Attr.TEX0MTXIDX + i].type = GX.AttrType.NONE;
-                }
-
-                if (fields.hasJoints && matrixIdxThing >= 2) {
-                    vcd[GX.Attr.PNMTXIDX].type = GX.AttrType.DIRECT;
-                    let texmtxNum = 0;
-                    // FIXME: what is this?
-                    // if (shaders[curShader].hasTexmtx01) {
-                    //     vcd[GX.Attr.TEX0MTXIDX + texmtxNum].type = GX.AttrType.DIRECT; 
-                    //     texmtxNum++;
-                    //     vcd[GX.Attr.TEX0MTXIDX + texmtxNum].type = GX.AttrType.DIRECT;
-                    //     texmtxNum++;
-                    // }
-                    vcd[GX.Attr.TEX0MTXIDX + texmtxNum].type = GX.AttrType.DIRECT;
-                    texmtxNum++;
-                    for (let i = 0; i < texMtxCount; i++) {
-                        vcd[GX.Attr.TEX0MTXIDX + texmtxNum].type = GX.AttrType.DIRECT;
-                        texmtxNum++;
-                    }
-                }
-
-                const posDesc = bits.get(1);
-                vcd[GX.Attr.POS].type = posDesc ? GX.AttrType.INDEX16 : GX.AttrType.INDEX8;
-
-                if (fields.hasNormals && (curShader.attrFlags & 1) != 0) {
-                    const nrmDesc = bits.get(1);
-                    if ((nrmTypeFlags & 8) != 0) {
-                        vcd[GX.Attr.NRM].type = GX.AttrType.NONE;
-                        vcd[GX.Attr.NBT].type = nrmDesc ? GX.AttrType.INDEX16 : GX.AttrType.INDEX8;
-                    } else {
-                        vcd[GX.Attr.NBT].type = GX.AttrType.NONE;
-                        vcd[GX.Attr.NRM].type = nrmDesc ? GX.AttrType.INDEX16 : GX.AttrType.INDEX8;
-                    }
-                } else {
-                    vcd[GX.Attr.NRM].type = GX.AttrType.NONE;
-                }
-
-                if ((curShader.attrFlags & 2) != 0) {
-                    const clr0Desc = bits.get(1);
-                    vcd[GX.Attr.CLR0].type = clr0Desc ? GX.AttrType.INDEX16 : GX.AttrType.INDEX8;
-                } else {
-                    vcd[GX.Attr.CLR0].type = GX.AttrType.NONE;
-                }
-
-                const texCoordDesc = bits.get(1);
-                if (curShader.hasTexCoord[0]) {
-                    // Note: texCoordDesc applies to all texture coordinates in the vertex
-                    for (let t = 0; t < 8; t++) {
-                        if (curShader.hasTexCoord[t]) {
-                            vcd[GX.Attr.TEX0 + t].type = texCoordDesc ? GX.AttrType.INDEX16 : GX.AttrType.INDEX8;
-                        } else {
-                            vcd[GX.Attr.TEX0 + t].type = GX.AttrType.NONE;
-                        }
-                    }
-                }
-                break;
-            case 4: // Set weights (skipped by SFA block renderer)
-                const numWeights = bits.get(4);
-                for (let i = 0; i < numWeights; i++) {
-                    bits.get(8);
-                }
-                break;
-            case 5: // End
-                done = true;
-                break;
-            default:
-                console.warn(`Skipping unknown model bits opcode ${opcode}`);
-                break;
+        runBitstream(bitsOffsets[0]); // Opaques
+        for (let i = 1; i < bitsOffsets.length; i++) {
+            if (bitsOffsets[i] != 0) {
+                runBitstream(bitsOffsets[i]); // Translucents and waters
             }
         }
     }
